feat: add overlay controller and example
