+++ lib/flutter_micro_app.dart
@@ -2,6 +2,7 @@ library flutter_micro_app;
 
 export 'src/controllers/app_event/micro_app_event_controller.dart';
 export 'src/controllers/navigators/navigator_controller.dart';
+export 'src/controllers/overlay/micro_overlay_controller.dart';
 export 'src/core/micro_app.dart';
 export 'src/core/micro_host.dart';
 export 'src/core/micro_page.dart';
@@ -13,5 +14,4 @@ export 'src/entities/router/base_route.dart';
 export 'src/presentation/widgets/micro_widget_builder.dart';
 export 'src/utils/constants/constants.dart';
 export 'src/utils/constants/path_separator.dart';
-export 'src/presentation/overlay/micro_overlay_controller.dart';
 export 'src/utils/typedefs.dart';
